✨ feat: 헬스장 등록 유무에 따른 화면 분기 처리
